Move some code to catkit.util etc
